# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dental Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Users
...
  /revisits:
    get:
      operationId: listRevisits
      tags: [revisits]
      summary: List revisits for an OP record
      parameters:
        - name: opRecordId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of revisits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Revisit"
    post:
      operationId: createRevisit
      tags: [revisits]
      summary: Create a new revisit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revisit"

  # Diagnoses
  /diagnoses:
    get:
      operationId: listDiagnoses
      tags: [diagnoses]
      summary: List diagnoses for a patient
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: opRecordId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of diagnoses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Diagnosis"
    post:
      operationId: createDiagnosis
      tags: [diagnoses]
      summary: Add a diagnosis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diagnosis"
  /diagnoses/{id}:
    patch:
      operationId: updateDiagnosis
      tags: [diagnoses]
      summary: Update a diagnosis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diagnosis"

  # Prescriptions
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions for a patient
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Add a prescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    patch:
      operationId: updatePrescription
      tags: [prescriptions]
      summary: Update a prescription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
    delete:
      operationId: deletePrescription
      tags: [prescriptions]
      summary: Delete a prescription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Treatments
  /treatments:
    get:
      operationId: listTreatments
      tags: [treatments]
      summary: List treatments for a patient
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of treatments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treatment"
    post:
      operationId: createTreatment
      tags: [treatments]
      summary: Add a treatment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreatmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
  /treatments/{id}:
    patch:
      operationId: updateTreatment
      tags: [treatments]
      summary: Update a treatment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreatmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"

  # Invoices
...
  /followups:
    get:
      operationId: listFollowups
      tags: [followups]
      summary: List follow-ups
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of follow-ups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
    post:
      operationId: createFollowup
      tags: [followups]
      summary: Schedule a follow-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowUpInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /followups/{id}:
    patch:
      operationId: updateFollowup
      tags: [followups]
      summary: Update follow-up status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowUpUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"

  # Dashboard
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role, status]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [administrator, receptionist, doctor]
        status:
          type: string
          enum: [active, inactive]
        lastLoginAt:
          type: ["string", "null"]
...
    Revisit:
      type: object
      required: [id, opRecordId, visitDate, createdAt]
      properties:
        id:
          type: integer
        opRecordId:
          type: integer
        visitDate:
          type: string
        notes:
          type: ["string", "null"]
        doctorId:
          type: ["integer", "null"]
        doctorName:
          type: ["string", "null"]
        createdAt:
          type: string
    RevisitInput:
      type: object
      required: [opRecordId, visitDate]
      properties:
        opRecordId:
          type: integer
        visitDate:
          type: string
        notes:
          type: string
        doctorId:
          type: integer
    Diagnosis:
      type: object
      required: [id, patientId, notes, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        opRecordId:
          type: ["integer", "null"]
        notes:
          type: string
        createdBy:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
    DiagnosisInput:
      type: object
      required: [patientId, notes]
      properties:
        patientId:
          type: integer
        opRecordId:
          type: integer
        notes:
          type: string
    DiagnosisUpdate:
      type: object
      properties:
        notes:
          type: string
    Prescription:
      type: object
      required: [id, patientId, medicineName, dosage, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        medicineName:
          type: string
        dosage:
          type: string
        instructions:
          type: ["string", "null"]
        recommendations:
          type: ["string", "null"]
        createdAt:
          type: string
    PrescriptionInput:
      type: object
      required: [patientId, medicineName, dosage]
      properties:
        patientId:
          type: integer
        medicineName:
          type: string
        dosage:
          type: string
        instructions:
          type: string
        recommendations:
          type: string
    PrescriptionUpdate:
      type: object
      properties:
        medicineName:
          type: string
        dosage:
          type: string
        instructions:
          type: string
        recommendations:
          type: string
    Treatment:
      type: object
      required: [id, patientId, procedure, status, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        toothNumber:
          type: ["string", "null"]
        procedure:
          type: string
        status:
          type: string
          enum: [planned, in-progress, completed, cancelled]
        cost:
          type: ["number", "null"]
        timelineOrder:
          type: ["integer", "null"]
        createdAt:
          type: string
    TreatmentInput:
      type: object
      required: [patientId, procedure, status]
      properties:
        patientId:
          type: integer
        toothNumber:
          type: string
        procedure:
          type: string
        status:
          type: string
          enum: [planned, in-progress, completed, cancelled]
        cost:
          type: number
        timelineOrder:
          type: integer
    TreatmentUpdate:
      type: object
      properties:
        toothNumber:
          type: string
        procedure:
          type: string
        status:
          type: string
          enum: [planned, in-progress, completed, cancelled]
        cost:
          type: number
...
    FollowUp:
      type: object
      required: [id, patientId, scheduledDate, status, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        scheduledDate:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    FollowUpInput:
      type: object
      required: [patientId, scheduledDate]
      properties:
        patientId:
          type: integer
        scheduledDate:
          type: string
        notes:
          type: string
    FollowUpUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, completed, cancelled]
        notes:
          type: string
        scheduledDate:
          type: string
...