# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Prescription {
  id: number;
  patientId: number;
  medicineName: string;
  dosage: string;
  /** @nullable */
  instructions?: string | null;
  /** @nullable */
  recommendations?: string | null;
  createdAt: string;
}