# lib/api-zod/src/generated/types/revisit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Revisit {
  id: number;
  opRecordId: number;
  visitDate: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  doctorId?: number | null;
  /** @nullable */
  doctorName?: string | null;
  createdAt: string;
}