# lib/api-zod/src/generated/types/treatmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TreatmentInputStatus } from './treatmentInputStatus';

export interface TreatmentInput {
  patientId: number;
  toothNumber?: string;
  procedure: string;
  status: TreatmentInputStatus;
  cost?: number;
  timelineOrder?: number;
}