# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  username: string;
  fullName?: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  lastLoginAt?: string | null;
}